fix some bugs add some refactoring code

+++ backend/src/main/java/ru/trainee/adminPanel/repository/BannerRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 import ru.trainee.adminPanel.model.Banner;
 
-public interface BannerRepository extends JpaRepository<Banner, Integer> {
+public interface BannerRepository extends JpaRepository<Banner, Long> {
     @Transactional
     @Modifying
     @Query(value = "insert into banner (banner_id, imgsrc, width, height, targeturl, langid) VALUES (?1, ?2, ?3, ?4, ?5, ?6)", nativeQuery = true)
-    int saveBanner(int banner_id, String imgsrc, int width, int height, String targeturl, String langid);
+    int saveBanner(Long banner_id, String imgsrc, int width, int height, String targeturl, String langid);
 }
